Map voting working with fontawesome
