+++ client/package.json
@@ -7,6 +7,7 @@
     "start": "sirv public"
   },
   "devDependencies": {
+    "@fortawesome/fontawesome-free": "^5.15.1",
     "@rollup/plugin-commonjs": "^14.0.0",
     "@rollup/plugin-node-resolve": "^8.0.0",
     "node-sass": "^4.14.1",
