+++ client/public/index.html
@@ -4,7 +4,6 @@
 	<meta charset='utf-8'>
 	<meta name='viewport' content='width=device-width,initial-scale=1'>
 
-    <link href="/css/all.min.css" rel="stylesheet">
     <link rel="shortcut icon" href="/favicon.ico">
     <title>SRB2 circuit race</title>
