+++ client/src/App.scss
@@ -3,7 +3,3 @@
 // MDC Typography
 //
 @import "@material/typography/mdc-typography";
-
-.container {
-  padding: 30px;
-}
