+++ client/src/api.js
@@ -1,0 +1,9 @@
+import {api_url} from './config.js';
+
+export function get_maps(callback) {
+    var url = new URL(api_url + '/maps')
+    return fetch(url)
+        .then(response => {
+            return response.json()
+        })
+}
