+++ client/src/main.js
@@ -1,5 +1,6 @@
 import './App.scss';
 import App from './App.svelte';
+import "@fortawesome/fontawesome-free/js/all";
 
 const app = new App({
 	target: document.body,
