+++ map_voting.py
@@ -4,12 +4,6 @@ from config import Config
 
 map_voting = Blueprint('map_voting', __name__)
 
-@map_voting.route('/')
-def list():
-    query = db.session.query(Map).order_by(Map.votes.desc()).filter(Map.in_rotation)
-    maps = query.all()
-    return render_template('map_voting.html', maps=maps, config=Config)
-
 @map_voting.route('/vote', methods = ['POST'])
 def vote():
     map = int(request.form['map'])
